# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for GitEcho, a tool for backing up Git repositories",
        "title": "GitEcho API",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "http://www.opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/repository": {
            "get": {
                "description": "Get a list of all backup repositories",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "backup-repositories"
                ],
                "summary": "Get all backup repositories",
                "responses": {
                    "200": {
                        "description": "List of backup repositories",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/backuprepo.BackupRepo"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new backup repository configuration with the given data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "backup-repositories"
                ],
                "summary": "Create a new backup repository configuration",
                "parameters": [
                    {
                        "description": "Backup repository data",
                        "name": "backupRepo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backuprepo.BackupRepo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response\"“",
                        "schema": {
                            "$ref": "#/definitions/server.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/repository/{repo_name}": {
            "get": {
                "description": "Get the backup repository with the given name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "backup-repositories"
                ],
                "summary": "Get a backup repository by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name of the backup repository to retrieve",
                        "name": "repo_name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Backup repository data",
                        "schema": {
                            "$ref": "#/definitions/backuprepo.BackupRepo"
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Backup repository not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a backup repository by its name",
                "tags": [
                    "backup-repositories"
                ],
                "summary": "Delete a backup repository",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name of the backup repository to delete",
                        "name": "repo_name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response",
                        "schema": {
                            "$ref": "#/definitions/server.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/repository/{repo_name}/storage/": {
            "get": {
                "description": "Get all storages associated with a backup repository by its name",
                "tags": [
                    "backup-repositories"
                ],
                "summary": "Get backup repository storages",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name of the backup repository",
                        "name": "repo_name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of storages associated with the backup repository",
                        "schema": {
                            "type": "array",
                            "items": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/repository/{repo_name}/storage/{storage_name}": {
            "post": {
                "description": "Associate a storage with a backup repository by their names",
                "tags": [
                    "backup-repositories"
                ],
                "summary": "Add storage to backup repository",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name of the backup repository",
                        "name": "repo_name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Name of the storage",
                        "name": "storage_name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response",
                        "schema": {
                            "$ref": "#/definitions/server.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove the association between a storage and a backup repository by their names",
                "tags": [
                    "backup-repositories"
                ],
                "summary": "Remove storage from backup repository",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name of the backup repository",
                        "name": "repo_name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Name of the storage",
                        "name": "storage_name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response",
                        "schema": {
                            "$ref": "#/definitions/server.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "backuprepo.BackupRepo": {
            "type": "object",
            "properties": {
                "credentials": {
                    "$ref": "#/definitions/backuprepo.Credentials"
                },
                "name": {
                    "type": "string"
                },
                "pull_interval": {
                    "type": "integer"
                },
                "remote_url": {
                    "type": "string"
                }
            }
        },
        "backuprepo.Credentials": {
            "type": "object",
            "properties": {
                "key_path": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "server.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}